Rename AlphaEquivalence to LemmaEquivalence
